requirements.txt
asgiref==3.8.1
cramjam==2.9.0
dataclasses==0.6
Django==5.1.3
djongo==1.2.31
dnspython==2.7.0
PyJWT==2.10.0
pykerberos==1.2.4
pymongo==4.10.1
python-decouple==3.8
python-dotenv==1.0.1
python-snappy==0.7.3
pytz==2024.2
sqlparse==0.5.2
